move expandoButton code to separate file
